mv storage to syncer/servicecenter
